Moved query api definition into separate module
